+++ src/components/formStepTwo/ProfilePicture.css
@@ -1,14 +1,64 @@
 .profile-picture-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .profile-picture-content {
-    display: flex;
-    align-items: center;
-  }
-  
-  .profile-picture-img {
-    max-width: 200px;
-  }
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: relative;
+}
+
+.profile-picture-content {
+  display: flex;
+  align-items: center;
+  position: relative;
+  overflow: hidden;
+  border-radius: 50%;
+  width: 150px;
+  height: 150px;
+  border: 3px solid #ed2590; 
+  margin-top: 10px;
+  background-color: transparent; /* cor de fundo transparente */
+  padding: 3px; /* 2px entre a imagem e a borda */
+}
+
+.profile-picture-img {
+  max-width: 100%;
+  max-height: 100%;
+  width: 100%;
+  height: 100%;
+  object-fit: cover; /* preencher todo o círculo */
+  object-position: center; /* centralizar */
+  border-radius: 50%; /* círculo */
+}
+
+.profile-picture-img::before {
+  content: "";
+  display: block;
+  padding-top: 100%; /* espaço igual ao raio do círculo */
+}
+
+input[type="file"] {
+  position: absolute;
+  top: 0;
+  left: 0;
+  opacity: 0;
+  width: 100%;
+  height: 100%;
+  cursor: pointer;
+}
+
+.photo-button{
+  background-color: black #FDFEFF;
+  border: 1px solid #ed2590; /* linha rosa ao redor do textarea */
+  border-radius: 5px;
+  color: #ed2590;
+  padding: 10px 20px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  margin-top: 10px;
+  cursor: pointer;
+}
+
+.profile-picture-text {
+  color: #4B4B4B
+}
