+++ src/components/formStepTwo/SuccessMessage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const SuccessMessage: React.FC = () => {
   return (
-    <div style={{ color: 'green', fontWeight: 'bold', fontSize: '18px' }}>
+    <div style={{ color: 'gray', fontWeight: 'bold', fontSize: '18px' }}>
       Cadastro realizado com sucesso!
     </div>
   );
