+++ src/components/formStepTwo/TextArea.css
@@ -1,0 +1,31 @@
+
+.textarea-container {
+    position: relative;
+    margin-bottom: 0px;
+    margin-top: 40px;
+  }
+  
+  .textarea-label {
+    position: absolute;
+    top: -20px; 
+    left: 10px; 
+    background-color: #FFFFFF;
+    padding: 0 2px;
+    pointer-events: none;
+    font-size: 15px;
+    transition: 0.2s;
+  }
+  
+  .textarea-input {
+    resize: none;
+    width: 100%;
+    padding: 20px;
+    font-size: 16px;
+    border: 1px solid #ed2590; 
+    border-radius: 5px;
+  }
+  
+  .textarea-container {
+    display: flex; 
+    align-items: center; 
+  }
